tracing_bug_bash/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
opentelemetry-sdk==1.19.0
opentelemetry-api==1.19.0
opentelemetry-exporter-otlp==1.19.0
opentelemetry-instrumentation-fastapi==0.40b0
requests==2.31.0
httpx==0.24.1
# For testing
pytest==8.0.0 